+++ src/main/java/com/coffee/demoCoffeeApp/repository/ProductoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.coffee.demoCoffeeApp.model.ProductoModel;
 
-public interface ProductoRepository extends JpaRepository<ProductoModel,Integer>{
+public interface ProductoRepository extends JpaRepository<ProductoModel,Long>{
     
 }
